+++ packages/web3-utils/CHANGELOG.md
@@ -83,3 +83,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Added support of `safe` and `finalized` block tags (#5823)
 
+### Changed
+
+-   `compareBlockNumbers` function now only supports comparison of both blocktags params ( except `earliest` vs number) or both block number params (#5842)